5398: abstract sections into files
