+++ src/apps/patron-page/PatronPage.mount.ts
@@ -1,0 +1,4 @@
+import addMount from "../../core/addMount";
+import PatronPage from "./PatronPage.entry";
+
+addMount({ appName: "patron-page", app: PatronPage });
